Main class' method now consists of creating a MainMenu instance and then calling its .display() and .getAndExecuteAction() methods in a while loop.
Added a .login() method to the UAC class for use by MainMenu objects, upon their construction.

Also moved all the book-related methods to the BookUtils class for use by MainMenu objects.
The public methods use the private utility methods in the class (e.g: addBook() is to be called externally by a MainMenu object, and it uses the private initialiseFile(),writeToFile() and getBookInfo() methods)